chore(clerk-js): Remove barrel files
