+++ packages/clerk-js/src/ui/components/SignIn/__tests__/ResetPasswordSuccess.test.tsx
@@ -1,6 +1,7 @@
 import { describe, it } from '@jest/globals';
+import { screen } from '@testing-library/react';
 
-import { render, screen } from '../../../../testUtils';
+import { render } from '../../../../testUtils';
 import { bindCreateFixtures } from '../../../utils/test/createFixtures';
 import { runFakeTimers } from '../../../utils/test/runFakeTimers';
 import { ResetPasswordSuccess } from '../ResetPasswordSuccess';
