+++ packages/clerk-js/src/ui/components/UserProfile/__tests__/ConnectedAccountsSection.test.tsx
@@ -1,8 +1,8 @@
 import type { ExternalAccountResource } from '@clerk/types';
 import { describe, it } from '@jest/globals';
-import { act, waitFor } from '@testing-library/react';
+import { act, screen, waitFor } from '@testing-library/react';
 
-import { render, screen } from '../../../../testUtils';
+import { render } from '../../../../testUtils';
 import { bindCreateFixtures } from '../../../utils/test/createFixtures';
 import { ConnectedAccountsSection } from '../ConnectedAccountsSection';
