+++ packages/clerk-js/src/ui/components/UserProfile/__tests__/PasskeysSection.test.tsx
@@ -1,8 +1,8 @@
 import type { PasskeyJSON, PasskeyResource } from '@clerk/types';
 import { describe, it } from '@jest/globals';
-import { act } from '@testing-library/react';
+import { act, waitFor } from '@testing-library/react';
 
-import { render, waitFor } from '../../../../testUtils';
+import { render } from '../../../../testUtils';
 import { CardStateProvider } from '../../../elements';
 import { bindCreateFixtures } from '../../../utils/test/createFixtures';
 import { PasskeySection } from '../PasskeySection';
