+++ packages/clerk-js/src/ui/components/UserProfile/__tests__/UserProfileSection.test.tsx
@@ -1,7 +1,8 @@
 import type { ImageResource } from '@clerk/types';
 import { describe, it } from '@jest/globals';
+import { screen, waitFor } from '@testing-library/react';
 
-import { render, screen, waitFor } from '../../../../testUtils';
+import { render } from '../../../../testUtils';
 import { bindCreateFixtures } from '../../../utils/test/createFixtures';
 import { UserProfileSection } from '../UserProfileSection';
