+++ packages/clerk-js/src/ui/components/Waitlist/__tests__/Waitlist.test.tsx
@@ -1,6 +1,7 @@
 import type { WaitlistResource } from '@clerk/types';
+import { screen } from '@testing-library/react';
 
-import { render, screen } from '../../../../testUtils';
+import { render } from '../../../../testUtils';
 import { bindCreateFixtures } from '../../../utils/test/createFixtures';
 import { Waitlist } from '../Waitlist';
