+++ packages/clerk-js/src/ui/hooks/__tests__/useCoreOrganization.test.tsx
@@ -1,7 +1,7 @@
 import { useOrganization } from '@clerk/shared/react';
 import { describe } from '@jest/globals';
+import { act, renderHook, waitFor } from '@testing-library/react';
 
-import { act, renderHook, waitFor } from '../../../testUtils';
 import {
   createFakeDomain,
   createFakeOrganizationInvitation,
