+++ packages/clerk-js/src/ui/primitives/FormErrorText.tsx
@@ -4,7 +4,7 @@ import { Icon } from '../customizables';
 import { ExclamationCircle } from '../icons';
 import type { StyleVariants } from '../styledSystem';
 import { animations, createVariants } from '../styledSystem';
-import { useFormField } from './hooks';
+import { useFormField } from './hooks/useFormField';
 import { Text } from './Text';
 
 const { applyVariants } = createVariants(theme => ({
