+++ packages/clerk-js/src/ui/primitives/FormInfoText.tsx
@@ -2,7 +2,7 @@ import { forwardRef } from 'react';
 
 import type { FormTextProps } from './FormSuccessText';
 import { applyVariants } from './FormSuccessText';
-import { useFormField } from './hooks';
+import { useFormField } from './hooks/useFormField';
 import { Text } from './Text';
 
 export const FormInfoText = forwardRef<HTMLElement, FormTextProps>((props, ref) => {
