--- packages/clerk-js/src/ui/primitives/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useFormField';
-export * from './useInput';
