--- packages/clerk-js/src/utils/captcha/index.ts
@@ -1,3 +1,0 @@
-export * from './retrieveCaptchaInfo';
-export * from './constants';
-export * from './getCaptchaToken';
